+++ lib/multi_qr_tracker.dart
@@ -20,6 +20,7 @@ export 'src/models/qr_code_info.dart';
 export 'src/models/qr_tracker_camera_orientation.dart';
 export 'src/models/torch_button_position.dart';
 export 'src/models/torch_mode.dart';
+export 'src/multi_qr_tracker_controller.dart';
 export 'src/multi_qr_tracker_view.dart';
 export 'src/widgets/qr_border_painter.dart';
 export 'src/widgets/torch_button.dart';